requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
appdirs==1.4.4
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
blinker==1.4
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
cytoolz==0.12.0
dask==2023.4.1
debugpy==1.5.1
decorator==5.1.1
executing==0.8.3
flatbuffers==2.0
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2023.4.0
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.7.0
idna==3.4
imagecodecs==2021.8.26
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.2.0
ipykernel==6.15.0
ipython==8.12.0
jedi==0.18.1
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
multidict==6.0.2
munkres==1.1.4
nest-asyncio==1.5.6
networkx==2.8.4
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
partd==1.2.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==2.5.2
ply==3.11
pooch==1.4.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.1
PyJWT==2.4.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.4.1
pywin32==305.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.29.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
setuptools==67.8.0
sip==6.6.2
six==1.16.0
stack-data==0.2.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
termcolor==2.1.0
threadpoolctl==2.2.0
tifffile==2021.7.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.7.1
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.5
Werkzeug==2.2.3
wheel==0.38.4
win-inet-pton==1.1.0
wrapt==1.14.1
yarl==1.8.1
zipp==3.11.0